requirements.txt
streamlit
pandas
psycopg2-binary
anthropic
python-dotenv
PyPDF2